Consolidate the pipeline stage classes
