+++ Microsoft.Azure.Cosmos/src/Query/Core/Pipeline/Aggregate/AggregateQueryPipelineStage.cs
@@ -23,96 +23,5 @@ namespace Microsoft.Azure.Cosmos.Query.Core.Pipeline.Aggregate
     /// </summary>
     internal abstract partial class AggregateQueryPipelineStage : QueryPipelineStageBase
     {
-        /// <summary>
-        /// This class does most of the work, since a query like:
-        /// 
-        /// SELECT VALUE AVG(c.age)
-        /// FROM c
-        /// 
-        /// is really just an aggregation on a single grouping (the whole collection).
-        /// </summary>
-        private readonly SingleGroupAggregator singleGroupAggregator;
-
-        /// <summary>
-        /// We need to keep track of whether the projection has the 'VALUE' keyword.
-        /// </summary>
-        private readonly bool isValueQuery;
-
-        protected bool returnedFinalPage;
-
-        /// <summary>
-        /// Initializes a new instance of the AggregateDocumentQueryExecutionComponent class.
-        /// </summary>
-        /// <param name="source">The source component that will supply the local aggregates from multiple continuations and partitions.</param>
-        /// <param name="singleGroupAggregator">The single group aggregator that we will feed results into.</param>
-        /// <param name="isValueQuery">Whether or not the query has the 'VALUE' keyword.</param>
-        /// <remarks>This constructor is private since there is some async initialization that needs to happen in CreateAsync().</remarks>
-        public AggregateQueryPipelineStage(
-            IQueryPipelineStage source,
-            SingleGroupAggregator singleGroupAggregator,
-            bool isValueQuery)
-            : base(source)
-        {
-            this.singleGroupAggregator = singleGroupAggregator ?? throw new ArgumentNullException(nameof(singleGroupAggregator));
-            this.isValueQuery = isValueQuery;
-        }
-
-        public static TryCatch<IQueryPipelineStage> MonadicCreate(
-            IReadOnlyList<AggregateOperator> aggregates,
-            IReadOnlyDictionary<string, AggregateOperator?> aliasToAggregateType,
-            IReadOnlyList<string> orderedAliases,
-            bool hasSelectValue,
-            CosmosElement continuationToken,
-            MonadicCreatePipelineStage monadicCreatePipelineStage)
-        {
-            return ClientAggregateQueryPipelineStage.MonadicCreate(
-                    aggregates,
-                    aliasToAggregateType,
-                    orderedAliases,
-                    hasSelectValue,
-                    continuationToken,
-                    monadicCreatePipelineStage);
-        }
-
-        /// <summary>
-        /// Struct for getting the payload out of the rewritten projection.
-        /// </summary>
-        private readonly struct RewrittenAggregateProjections
-        {
-            public RewrittenAggregateProjections(bool isValueAggregateQuery, CosmosElement raw)
-            {
-                if (raw == null)
-                {
-                    throw new ArgumentNullException(nameof(raw));
-                }
-
-                if (isValueAggregateQuery)
-                {
-                    // SELECT VALUE [{"item": {"sum": SUM(c.blah), "count": COUNT(c.blah)}}]
-                    if (!(raw is CosmosArray aggregates))
-                    {
-                        throw new ArgumentException($"{nameof(RewrittenAggregateProjections)} was not an array for a value aggregate query. Type is: {raw.GetType()}");
-                    }
-
-                    this.Payload = aggregates[0];
-                }
-                else
-                {
-                    if (!(raw is CosmosObject cosmosObject))
-                    {
-                        throw new ArgumentException($"{nameof(raw)} must not be an object.");
-                    }
-
-                    if (!cosmosObject.TryGetValue("payload", out CosmosElement cosmosPayload))
-                    {
-                        throw new InvalidOperationException($"Underlying object does not have an 'payload' field.");
-                    }
-
-                    this.Payload = cosmosPayload ?? throw new ArgumentException($"{nameof(RewrittenAggregateProjections)} does not have a 'payload' property.");
-                }
-            }
-
-            public CosmosElement Payload { get; }
-        }
     }
 }
