+++ Microsoft.Azure.Cosmos/src/Query/Core/Pipeline/DCount/DCountQueryPipelineStage.Client.cs
@@ -6,20 +6,70 @@ namespace Microsoft.Azure.Cosmos.Query.Core.Pipeline.DCount
 {
     using System;
     using System.Collections.Generic;
-    using System.Collections.Immutable;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.CosmosElements;
     using Microsoft.Azure.Cosmos.CosmosElements.Numbers;
     using Microsoft.Azure.Cosmos.Query.Core.Monads;
-    using Microsoft.Azure.Cosmos.Query.Core.Pipeline.Aggregate;
-    using Microsoft.Azure.Cosmos.Query.Core.Pipeline.Aggregate.Aggregators;
     using Microsoft.Azure.Cosmos.Query.Core.Pipeline.Pagination;
     using Microsoft.Azure.Cosmos.Query.Core.QueryPlan;
     using Microsoft.Azure.Cosmos.Tracing;
 
+    /// <summary>
+    /// Stage that is able to aggregate COUNT(DISTINCT) from multiple continuations and partitions.
+    /// </summary>
     internal abstract partial class DCountQueryPipelineStage : QueryPipelineStageBase
     {
+                /// <summary>
+        /// We need to keep track of whether the projection has the 'VALUE' keyword or an alias.
+        /// </summary>
+        private readonly DCountInfo info;
+
+        /// <summary>
+        /// This job of this class is to just keep a count.
+        /// </summary>
+        private long count;
+
+        protected bool returnedFinalPage;
+
+        /// <summary>
+        /// Initializes a new instance of the DCountQueryPipelineStage class.
+        /// </summary>
+        /// <param name="source">The source component that will supply the local aggregates from multiple continuations and partitions.</param>
+        /// <param name="count">The actual dcount that will be reported.</param>
+        /// <param name="info">Metadata about the original dcount query that is elided in the rewritten query</param>
+        /// <remarks>This constructor is private since there is some async initialization that needs to happen in CreateAsync().</remarks>
+        public DCountQueryPipelineStage(
+            IQueryPipelineStage source,
+            long count,
+            DCountInfo info)
+            : base(source)
+        {
+            this.count = count;
+            this.info = info;
+        }
+
+        public static TryCatch<IQueryPipelineStage> MonadicCreate(
+            DCountInfo info,
+            CosmosElement continuationToken,
+            MonadicCreatePipelineStage monadicCreatePipelineStage)
+        {
+            return ClientDCountQueryPipelineStage.MonadicCreate(
+                info,
+                continuationToken,
+                monadicCreatePipelineStage);
+        }
+
+        protected CosmosElement GetFinalResult()
+        {
+            return this.info.IsValueAggregate ?
+                CosmosNumber64.Create(this.count) as CosmosElement :
+                CosmosObject.Create(new Dictionary<string, CosmosElement>
+                {
+                    { this.info.DCountAlias, CosmosNumber64.Create(this.count) }
+                });
+        }
+
         private sealed class ClientDCountQueryPipelineStage : DCountQueryPipelineStage
         {
             private ClientDCountQueryPipelineStage(
