+++ Microsoft.Azure.Cosmos/src/Query/Core/Pipeline/Distinct/DistinctQueryPipelineStage.cs
@@ -21,28 +21,5 @@ namespace Microsoft.Azure.Cosmos.Query.Core.Pipeline.Distinct
     /// </summary>
     internal abstract partial class DistinctQueryPipelineStage : QueryPipelineStageBase
     {
-        /// <summary>
-        /// An DistinctMap that efficiently stores the documents that we have already seen.
-        /// </summary>
-        private readonly DistinctMap distinctMap;
-
-        protected DistinctQueryPipelineStage(
-            DistinctMap distinctMap,
-            IQueryPipelineStage source)
-            : base(source)
-        {
-            this.distinctMap = distinctMap ?? throw new ArgumentNullException(nameof(distinctMap));
-        }
-
-        public static TryCatch<IQueryPipelineStage> MonadicCreate(
-            CosmosElement requestContinuation,
-            MonadicCreatePipelineStage monadicCreatePipelineStage,
-            DistinctQueryType distinctQueryType)
-        {
-            return ClientDistinctQueryPipelineStage.MonadicCreate(
-                requestContinuation,
-                monadicCreatePipelineStage,
-                distinctQueryType);
-        }
     }
 }
